+++ netflipster/csharp/Netflipster.tests/App.config
@@ -5,7 +5,7 @@
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
   <connectionStrings>
-    <add name="Netflix2" connectionString="Data Source=|DataDirectory|\dbtest.db" />
+    <add name="Netflipster" connectionString="Data Source=|DataDirectory|\dbtest.db" />
   </connectionStrings>
   <entityFramework>
     <defaultConnectionFactory type="System.Data.Entity.Infrastructure.LocalDbConnectionFactory, EntityFramework">
