+++ netflipster/csharp/Netflipster.tests/BookingDaoTests.cs
@@ -1,8 +1,8 @@
-﻿using Netflix2.repository;
+﻿using Netflipster.Repository;
 using NFluent;
 using NUnit.Framework;
 
-namespace Netflix2.tests
+namespace Netflipster.tests
 {
     [TestFixture]
     public class BookingDaoTests
