+++ netflipster/csharp/Netflipster.tests/UserDaoTests.cs
@@ -1,8 +1,8 @@
-﻿using Netflix2.repository;
+﻿using Netflipster.Repository;
 using NFluent;
 using NUnit.Framework;
 
-namespace Netflix2.tests
+namespace Netflipster.tests
 {
     [TestFixture]
     public class UserDaoTests
