+++ netflipster/csharp/Netflipster.tests/packages.config
