+++ netflipster/csharp/Netflipster/App.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <connectionStrings>
-    <add name="Netflix2" connectionString="Data Source=|DataDirectory|\Netflix2.db" />
+    <add name="Netflipster" connectionString="Data Source=|DataDirectory|\Netflipster.db" />
   </connectionStrings>
 
   <startup>
