+++ netflipster/csharp/Netflipster/Main.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text;
 
-namespace Netflix2
+namespace Netflipster
 {
     class Program
     {
