+++ netflipster/csharp/Netflipster/Netflipster.csproj
@@ -8,8 +8,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{4CAF137B-2F70-4DBC-BBF6-3AF284A3B49F}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>Netflix2</RootNamespace>
-    <AssemblyName>Netflix2</AssemblyName>
+    <RootNamespace>Netflipster</RootNamespace>
+    <AssemblyName>Netflipster</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -65,14 +65,14 @@
   <ItemGroup>
     <Compile Include="Main.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="repository\BookingDao.cs" />
-    <Compile Include="repository\DB.cs" />
-    <Compile Include="repository\UserDao.cs" />
-    <Compile Include="repository\VhsDao.cs" />
+    <Compile Include="Repository\BookingDao.cs" />
+    <Compile Include="Repository\DB.cs" />
+    <Compile Include="Repository\UserDao.cs" />
+    <Compile Include="Repository\VhsDao.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
-    <None Include="Netflix2.db">
+    <None Include="Netflipster.db">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Include="packages.config" />
