+++ netflipster/csharp/netflipster_csharp.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2018
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Netflipster", "Netflipster\Netflipster.csproj", "{4CAF137B-2F70-4DBC-BBF6-3AF284A3B49F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4CAF137B-2F70-4DBC-BBF6-3AF284A3B49F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CAF137B-2F70-4DBC-BBF6-3AF284A3B49F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CAF137B-2F70-4DBC-BBF6-3AF284A3B49F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CAF137B-2F70-4DBC-BBF6-3AF284A3B49F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4DC18E42-9328-4E51-96CE-1316A27945AC}
+	EndGlobalSection
+EndGlobal
